+++ common/src/main/java/de/cech12/ceramicshears/Constants.java
@@ -1,7 +1,10 @@
 package de.cech12.ceramicshears;
 
+import net.minecraft.world.item.Item;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.function.Supplier;
 
 /**
  * Class that contains all common constants.
@@ -14,6 +17,9 @@ public class Constants {
     public static final String MOD_NAME = "Ceramic Shears";
     /** Logger instance */
     public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
+
+    /** Supplier of registered ceramic shears item */
+    public static Supplier<Item> CERAMIC_SHEARS;
 
     private Constants() {}