+++ fabric/src/main/java/de/cech12/ceramicshears/FabricCeramicShearsMod.java
@@ -4,15 +4,18 @@ import de.cech12.ceramicshears.item.CeramicShearsItem;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.minecraft.core.Registry;
+import net.minecraft.core.dispenser.ShearsDispenseItemBehavior;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.DispenserBlock;
 
 /**
  * Mod class for the Fabric loader.
  */
-public class CeramicShearsMod implements ModInitializer {
+@SuppressWarnings("unused")
+public class FabricCeramicShearsMod implements ModInitializer {
 
     /** clay shears part item registry object */
     public static final Item CLAY_SHEARS_PART = Registry.register(BuiltInRegistries.ITEM, new ResourceLocation(Constants.MOD_ID, "clay_shears_part"), new Item(new Item.Properties()));
@@ -20,6 +23,10 @@ public class CeramicShearsMod implements ModInitializer {
     public static final Item CERAMIC_SHEARS_PART = Registry.register(BuiltInRegistries.ITEM, new ResourceLocation(Constants.MOD_ID, "ceramic_shears_part"), new Item(new Item.Properties()));
     /** ceramic shears item registry object */
     public static final Item CERAMIC_SHEARS = Registry.register(BuiltInRegistries.ITEM, new ResourceLocation(Constants.MOD_ID, "ceramic_shears"), new CeramicShearsItem());
+
+    static {
+        Constants.CERAMIC_SHEARS = () -> CERAMIC_SHEARS;
+    }
 
     /**
      * Initialization of a mod instance.
@@ -35,8 +42,8 @@ public class CeramicShearsMod implements ModInitializer {
         ItemGroupEvents.modifyEntriesEvent(CreativeModeTabs.TOOLS_AND_UTILITIES).register(content -> {
             content.accept(CERAMIC_SHEARS);
         });
-        //register dispense behavior (is done by wsmlmb)
-        //DispenserBlock.registerBehavior(CERAMIC_SHEARS, new ShearsDispenseItemBehavior());
+        //register dispense behavior
+        DispenserBlock.registerBehavior(CERAMIC_SHEARS, new ShearsDispenseItemBehavior());
     }
 
 }