+++ fabric/src/main/java/de/cech12/ceramicshears/mixin/EntityInteractShearsMixin.java
@@ -1,0 +1,25 @@
+package de.cech12.ceramicshears.mixin;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import de.cech12.ceramicshears.item.CeramicShearsItem;
+import net.minecraft.world.entity.animal.MushroomCow;
+import net.minecraft.world.entity.animal.Sheep;
+import net.minecraft.world.entity.animal.SnowGolem;
+import net.minecraft.world.entity.animal.Wolf;
+import net.minecraft.world.entity.monster.Bogged;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin({Bogged.class, MushroomCow.class, Sheep.class, SnowGolem.class, Wolf.class})
+public abstract class EntityInteractShearsMixin {
+
+	@WrapOperation(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/ItemStack;is(Lnet/minecraft/world/item/Item;)Z"), method = "mobInteract")
+	private boolean isShears(ItemStack stack, Item item, Operation<Boolean> original) {
+		return original.call(stack, item) || (item == Items.SHEARS && stack.getItem() instanceof CeramicShearsItem);
+	}
+
+}