+++ CHANGELOG.md
@@ -3,6 +3,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.20.6-2.3.0.0] - 2024-06-08
+### Changed
+- Updated to Minecraft 1.20.6 (Fabric 0.98.0+1.20.6, Neoforge 20.6.112-beta, Forge 50.1.0)
+- Updated Cloth Config support (14.0.126) (Fabric/Quilt)
+- Updated ModMenu support (10.0.0-beta.1) (Fabric/Quilt)
+
+### Removed
+- dependency WSMLMB is not needed any longer (Fabric/Quilt)
+- removed unused shears tags (Neoforge, Fabric/Quilt)
+
+### Known Issues
+- wolf armor cannot be sheared off (Neoforge, Forge)
+- Bogged cannot be sheared (Forge)
+
 ## [1.20.4-2.2.0.0] - 2024-04-13
 ### Added
 - add Fabric (>=0.96.11+1.20.4) support (Fabric, Quilt)